requirements.txt
Flask==3.0.0
joblib==1.4.0
numpy>=1.26.0
scikit-learn==1.3.2
Werkzeug==3.0.0
supabase==2.3.0
python-dotenv==1.0.0
